pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eqnorm"
version = "0.1.0"
description = "A machine learning potential energy project eqnorm"
authors = [
    {name = "Yuzhuo Chen", email = "chenyuzhuo@zhejianglab.org"},
    {name = "Lyuwen Fu", email = "lyuwenfu@zhejianglab.org"},
    {name = "Shuxiang Yang", email = "yang_shuxiang@zhejianglab.org"},
    {name = "Lipeng Chen", email = "chenlp@zhejianglab.org"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["machine learning", "potential energy", "eqnorm"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "ase",
    "yaml",
    "argparse",
    "importlib",
    "vesin",
    "torch_geometric",
    "setuptools>=61.0",
    "e3nn>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/yzchen08/eqnorm"
Repository = "https://github.com/yzchen08/eqnorm"

[tool.setuptools.packages.find]
include = ["eqnorm"]
exclude = ["log"]

[tool.setuptools.package-data]
"eqnorm" = ["model_settings/*"]